requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cssselect==1.3.0
dill==0.3.9
discord.py==2.6.3
Django==4.2.27
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.14.0

dnspython==2.7.0
exceptiongroup==1.3.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
flower==2.0.1
frozenlist==1.8.0
fsspec==2025.3.2
gunicorn==25.1.0
huggingface-hub==0.30.2
humanize==4.15.0
idna==3.10
isort==5.13.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
keybert==0.9.0
kombu==5.6.1
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.1

packaging==25.0
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.0
Pygments==2.19.2
pylint==3.3.3
pymongo==3.11.4
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
rake-nltk==1.0.6
redis==7.1.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.3.3
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentencepiece==0.2.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.7
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tomlkit==0.13.2

tornado==6.5.4
tqdm==4.67.1
transformers==4.51.3

typing_extensions==4.13.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.21.0
